+++ abstracraft-core/src/main/java/tools/redstone/abstracraft/analysis/Dependency.java
@@ -1,10 +1,10 @@
 package tools.redstone.abstracraft.analysis;
 
-import tools.redstone.abstracraft.AbstractionManager;
+import tools.redstone.abstracraft.AbstractionProvider;
 
 public interface Dependency {
 
     /** Check whether this dependency is fully implemented */
-    boolean isImplemented(AbstractionManager manager);
+    boolean isImplemented(AbstractionProvider manager);
 
 }
